backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
google-generativeai
requests
passlib[bcrypt]
python-jose[cryptography]
httpx
websockets